cosmocartt_Bot/backend/requirements.txt

fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pandas
pydantic
requests

# LangChain and LLM Integration
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-groq>=0.0.1

# Alternative LLM providers (install as needed):
# langchain-google-genai>=0.0.5  # For Google Gemini
# langchain-anthropic>=0.0.5     # For Anthropic Claude
# ollama>=0.1.0                   # For Local LLMs
